requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1 

faiss-cpu>=1.7.4 

pypdf>=3.17.0
PyMuPDF>=1.23.0  

rank-bm25>=0.2.2

requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

tqdm>=4.66.0

numpy>=1.24.0
pandas>=2.1.0

sentence-transformers>=2.2.2 
nltk>=3.8.1  

scikit-learn>=1.3.0  
ragas>=0.1.0  
datasets>=2.14.0